Add fix for gizmos by using relative quaternion.euler for all piece placment
